studio/video_remover/requirements.txt
moviepy==1.0.3
rembg
pillow
numpy
aiohttp
aiofiles
imageio<3.0.0
decorator<5.0.0
